requirements.txt
tensorflow==2.20.0
numpy>=1.24
pillow>=10.0
matplotlib>=3.7
scikit-learn>=1.3
Flask>=3.0
werkzeug>=3.0
opencv-python

